delete file from changelog
--HG--
branch : capi-sync-deltas


Former-commit-id: 627b7e81e9220d593d18bdc3b4e78cdd6c5115bd

+++ Mono/CAPI.Android.Model/ChangeLog/FileChangeLogStore.cs
@@ -25,7 +25,8 @@ namespace CAPI.Android.Core.Model.ChangeLog
         public void DeleteDraftChangeSet(Guid recordId)
         {
             var path = GetFileName(recordId);
-            File.Delete(path);
+            if (File.Exists(path))
+                File.Delete(path);
         }
 
         private string GetFileName(Guid publicKey)
